setup client routing
